requirements.txt
streamlit==1.32.2         # For building interactive UI
pymongo==4.6.1            # MongoDB database operations
pandas==2.2.2             # Data handling and exports
bcrypt==4.1.3             # Secure password hashing
fpdf==1.7.2               # PDF receipt generation
openpyxl==3.1.2           # Excel export support
xlsxwriter==3.1.9         # Alternative Excel writer with styling
python-dateutil==2.9.0    # Date parsing and validation
python-dotenv==1.0.0
twilio==9.6.5
